+++ Source/Nexus/NexusGameMode.h
@@ -1,0 +1,24 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameModeBase.h"
+#include "NexusGameMode.generated.h"
+
+/**
+ *  Simple GameMode for a third person game
+ */
+UCLASS(abstract)
+class ANexusGameMode : public AGameModeBase
+{
+	GENERATED_BODY()
+
+public:
+	
+	/** Constructor */
+	ANexusGameMode();
+};
+
+
+